% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_folder.R
\name{name_folder}
\alias{name_folder}
\title{Construct the standard GTS output folder name}
\usage{
name_folder(name, layer)
}
\arguments{
\item{name}{A character string used as the prefix (e.g. a site or project
name).}

\item{layer}{A character string giving the GTS layer short-code (e.g.
`"tm"`, `"rr"`). See \code{\link{VALID_GTS_LAYER_NAMES}}.}
}
\value{
A character string of the form `"<name>_<layer>"`.
}
\description{
Combines a user-supplied name and a GTS layer identifier into the folder
name convention used throughout the package (`<name>_<layer>`).
}
\examples{
name_folder("oslo_area", "tm")
#> [1] "oslo_area_tm"

}
\seealso{
\code{\link{download_nve_gts}},
  \code{\link{load_and_bind_gts_from_folder}},
  \code{\link{gts_folder_reset}}
}
